accidental redef of function f as variable

+++ src/fortran/fortran.c
@@ -491,7 +491,7 @@ FORTRANIFY(pshmem_int8_cswap)(long *target, long *cond, long *value, int *pe)
 
 
 void
-FORTRANIFY(pshmem_broadcast4)(void *target, const void *source, size_t *nelems,
+FORTRANIFY(pshmem_broadcast4)(void *target, const void *source, int *nelems,
 			      int *PE_root, int *PE_start,
 			      int *logPE_stride, int *PE_size,
 			      long *pSync)
@@ -513,7 +513,7 @@ FORTRANIFY(pshmem_broadcast8)(void *target, const void *source, size_t *nelems,
 }
 
 void
-FORTRANIFY(pshmem_broadcast32)(void *target, const void *source, size_t *nelems,
+FORTRANIFY(pshmem_broadcast32)(void *target, const void *source, int *nelems,
 			       int *PE_root, int *PE_start,
 			       int *logPE_stride, int *PE_size,
 			       long *pSync)
